requirements-core.txt
# 核心依赖 - 运行程序必需
# 数据处理
pandas==2.3.3
numpy==2.4.1

# 数据获取
requests==2.32.5

# 数据可视化
matplotlib==3.10.8
seaborn==0.13.2

# PDF生成
reportlab==4.2.5
Pillow==11.0.0

# 配置管理
PyYAML==6.0.2

# 日志
loguru==0.7.3
